feat: introduce process command for managing persistent processes
- Added a new `process` command to manage process runs, including subcommands for attaching, killing, listing, and spawning processes.
- Implemented `buildProcessCommand` in `process-command.ts` to handle command structure and functionality.
- Updated `composer-command.ts` to include the new process command and removed the deprecated runs command.
- Created a new `RunService` to encapsulate process management logic.
- Added end-to-end tests for the process command to ensure functionality and reliability.
- Refactored related tests to align with the new command structure.
